Add test for ostree-fuse
We try and recreate our test-data `fs.tar` by tarring up the contents of
the FUSE filesystem.  This makes it easy to test additional edge-cases as
all we have to do is add examples to `fs.tar`.

If it doesn't match the diffing is a little awkward.  We rely on
`diffoscope` to recursively diff inside the tarball we generate.  This
helps narrow down exactly what's gone wrong.

I'm leaning on tar in this instance as I don't want to implement my own
filesystem walking and comparison.  I'm hoping that tar is sufficiently
mature to get all the details right.  We can't extrace fs.tar and do a
comparison there as it contains root:root owned files and SUID bits, etc.

This test currently fails.