Repo: Replace `for_each_object` with `iter_objects`
This is a bit uglier on the implementation side, but it allows users to
take advantage of iterator combinators, etc.

This was much more difficult to implement than it sounds.  I had to patch
nix to have an owning variant of directory iteration to avoid
self-referential structs, and I've had to use nested `itertools::Either`
in `iter_objects_with_prefix` because even though we're using
`impl Iterator` as the return type, all the return sites must have the same
type.