Refactor: Add accessors to `struct Commit`
This is a bit nicer: parent returns `Option<CommitId>` which better
represents the underlying model.  We also have the option to change the
implementation in the future: I'd quite like to make it so `Commit`
contains a single `AlignedBytes<A8>` rather than each field individually,
but this will require more work in gvariant-rs to be efficient.

metadata and related objects are not currently exposed as the underlying
GVariant types are not currently nameable.  I've got some work in
gvariant-rs that will resolve this.