Add ostree-cmd prune
Attempts to replicate `ostree prune`.  This is entirely experimental at
this point and doesn't actually delete anything.

This uses rayon for parallelism.  It's significantly faster than ostree
prune in my testing, even with `RAYON_NUM_THREADS=1`.

It seems to give roughly equivalent results to `ostree prune` in terms
of number of objects to delete, but I've not written any proper tests
for it at this time.